requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-scheduler==2.13.3
google-auth==2.23.4
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.0
python-multipart==0.0.6
httpx==0.25.2
